30/04/21 сумма в поступлениях и правки по чеку

+++ app/Http/Controllers/api/v2/CronController.php
@@ -15,15 +15,19 @@ class CronController extends Controller
         $messages->each(function ($message) {
             try {
                 $order = Order::find($message['order_id']);
-                if ($order !== null) {
-                    $_message = (new CartController())->getMessage($order, null);
-                    \TelegramService::sendMessage($message['chat_id'], $_message);
-                    OrderMessage::find($message['id'])->update([
-                        'is_delivered' => true
-                    ]);
-                }
-                else {
-                    OrderMessage::find($message['id'])->delete();
+                if ($order['payment'] == Order::ORDER_PAYMENT_ONLINE && $order['is_paid'] == false) {
+                    return null;
+                } else {
+                    if ($order !== null) {
+                        $_message = (new CartController())->getMessage($order, null);
+                        \TelegramService::sendMessage($message['chat_id'], $_message);
+                        OrderMessage::find($message['id'])->update([
+                            'is_delivered' => true
+                        ]);
+                    }
+                    else {
+                        OrderMessage::find($message['id'])->delete();
+                    }
                 }
             } catch (\Exception $exception) {
                 throw $exception;
@@ -32,11 +36,11 @@ class CronController extends Controller
     }
 
     public function cancelOrders() {
-        /* $orders = Order::whereDate('created_at', '<=', now()->subDays(2))
-             ->whereDate('created_at', '>=', now()->subDays(7))
-             ->where('status', '==', 0)->get();
-         $orders->each(function ($order) {
-             (new OrderController())->decline($order);
-         });*/
+        $orders = Order::whereDate('created_at', '<=', now()->subDays(2))
+            ->whereDate('created_at', '>=', now()->subDays(7))
+            ->where('status', '==', 0)->get();
+        $orders->each(function ($order) {
+            (new OrderController())->decline($order);
+        });
     }
 }
